refactor(PUPIL-443): extract `InternalDndContext` and `usePrefersReducedMotion`
we'll need these for the match question
