+++ src/components/organisms/pupil/OakQuizOrder/OakQuizOrder.test.tsx
@@ -3,10 +3,11 @@ import { ThemeProvider } from "styled-components";
 import React from "react";
 import { DndContext, DndContextProps, UniqueIdentifier } from "@dnd-kit/core";
 
-import { OakQuizOrder, injectDndContext } from "./OakQuizOrder";
+import { OakQuizOrder } from "./OakQuizOrder";
 
 import { oakDefaultTheme } from "@/styles";
 import renderWithTheme from "@/test-helpers/renderWithTheme";
+import { injectDndContext } from "@/components/atoms/InternalDndContext/InternalDndContext";
 
 window.matchMedia = jest.fn().mockReturnValue({
   matches: false,
